# non_ascii_to_octal: Return the input string if it only contains printable ASCII characters